// src/app/shared/types/api-output-types/airworthiness-management/task-application-vo-output-dto.interface.ts
export interface TaskApplicationVOOutputDTO {
  taskCode?: string;
  taskVersion?: string;
  taskDescription?: string;
  taskType?: string;
  counterCode?: string;
  unitCode?: string;
  taskPeriodicity?: string;
  assetPn?: string;
  assetSn?: string;
  assetDescription?: string;
  assetChapter?: string;
  assetSection?: string;
  assetSubject?: string;
  assetSheet?: string;
  assetMarks?: string;
  assetItem?: string;
  ageing?: string;
  topAircraftMSN?: string;
  topAircraftRegistration?: string;
  nha?: string;
  nextTerm?: string;
  wpNumber?: string;
  woCode?: string;
  _expand?: boolean;
}