// src/app/shared/types/api-output-types/fleet-management/bido-equipment-ovs-output-dto.ts
export interface BidoEquipmentOVSOutputDTO {
  fatherEquipment?: string;
  equipmentCode?: string;
  familyVariantCode?: string;
  pnCode?: string;
  operatorCustomerCode?: string;
  ownerCustomerCode?: string;
  sn?: string;
  equipmentFunction?: string;
  equipmentFunctionCode?: string;
  equipmentDesignation?: string;
  registration?: string;
  modelCode?: string;
  operationalStatus?: string;
  ataCode?: string;
  batchNumber?: string;
}